chore: Refactor BusinessScatter component for improved performance and maintainability
