+++ src/app/(main-layout)/(protected)/spaces/[spaceId]/statistics/_components/report/report-charts/heat-map/index.ts
@@ -1,0 +1,1 @@
+export { default as HeatMap } from './heat-map';
